# Rewritten list query, test and fix [#82]