--- a/test/update_test.go
+++ b/test/update_test.go
@@ -157,16 +157,6 @@ func addSameEntries(ctx context.Context, g Gomega, client *clientv3.Client, numE
 	}
 }
 
-func addEntry(ctx context.Context, g Gomega, client *clientv3.Client, key string, value string) {
-	resp, err := client.Txn(ctx).
-		If(clientv3.Compare(clientv3.ModRevision(key), "=", 0)).
-		Then(clientv3.OpPut(key, value)).
-		Commit()
-
-	g.Expect(err).To(BeNil())
-	g.Expect(resp.Succeeded).To(BeTrue())
-}
-
 func updateEntry(ctx context.Context, g Gomega, client *clientv3.Client, key string, value string) {
 
 	resp, err := client.Get(ctx, key, clientv3.WithRange(""))