Solves [https://github.com/canonical/k8s-dqlite/issues/82](https://github.com/canonical/k8s-dqlite/issues/82)

Adds test to expose slowness of list query (https://github.com/canonical/k8s-dqlite/issues/82)
Original list query had maybe quadratic complexity [(plot)](https://www.wolframalpha.com/input?i=quadratic+fit+calculator&assumption=%7B%22F%22%2C+%22QuadraticFitCalculator%22%2C+%22data%22%7D+-%3E%22%7B2551219%2C4238600%2C6427871%2C9495735%2C14600343%2C18155273%2C24599527%2C29835900%2C38340280%2C48312167%2C56213177%2C67317845%2C69894101%2C88913819%2C102723989%2C142948612%2C124230579%2C143282009%2C169605043%7D%22) in terms of number of rows with same name.
New query passes the test, it  still has [superlinear](https://www.wolframalpha.com/input?i=linear+fit+calculator&assumption=%7B%22F%22%2C+%22LinearFitCalculator%22%2C+%22data%22%7D+-%3E%22%7B1948369%2C2440967%2C2547249%2C2096332%2C2208657%2C2143011%2C2219463%2C3453706%2C2177383%2C2983581%2C2320405%2C4230266%2C2220269%2C2279836%2C2210087%2C2294555%2C2515762%2C2526467%2C2786873%2C2557740%2C2663338%2C2712177%2C2600923%2C2615612%7D%22) complexity. But it has to be taken with grain of salt as I meassured individual updates.

Test performs two subsequents batches of 1000 updates.

Orignal query
- duration of first batch: 18 320 405 224ns
- duration of subsequent batch: 105 316 382 474ns

New Query
- duration of first batch:  2 372 000 975ns
- duration of subsequent batch duration: 2 684 759 515ns

Future work: By having specialized query for getting single name,(no prefix or range), we can have even faster, constant time complexity of list.