@ktsakalozos 2024-02-07
You can install MicroK8s with this patch with:
```
sudo snap install microk8s --classic --channel=latest/edge/dqlite-list
```