@miro-balaz 2024-02-20
Hallo, thanks for cooperation. I have some questions.
If I remember correctly I have signed CLA before (
https://github.com/canonical/microk8s/pull/2809). I hope it is still valid.

I am not sure what should I rebase and how. Do you mean to force rebase
branch in my fork?
something like this tutorial
https://github.com/AshleyGrant/git-tutorials/blob/master/Rebase%20Fork%20Branch%20to%20Upstream%20Master.md
?

   - git remote add upstream $URL where $URL is the HTTPS clone URL of the
   upstream repo
   - git pull --rebase upstream master
   - git checkout <featurebranch>
   - git rebase master
   - git push --force


ut 20. 2. 2024 o 6:39 Konstantinos Tsakalozos ***@***.***>
napísal(a):

> @miro-balaz <https://github.com/miro-balaz> thank you for this work we
> would like to merge it. Could you please rebase and sign the CLA so we can
> distribute the code? Thank you.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/canonical/k8s-dqlite/pull/83#issuecomment-1953520377>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AC3IHU5IUCU4BH35LAKTZT3YUQZI7AVCNFSM6AAAAABCX66XACVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTSNJTGUZDAMZXG4>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>