@miro-balaz 2024-02-26
Thanks, I would like to contribute more to this project, I already am
familiar with codebase. Is there any mailing list or slack? I would like to
know bit more about the future and why certain decisions were made.

st 21. 2. 2024 o 9:28 Berkay Tekin Öz ***@***.***> napísal(a):

> Merged #83 <https://github.com/canonical/k8s-dqlite/pull/83> into master.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/canonical/k8s-dqlite/pull/83#event-11872625362>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AC3IHU7HY3IHYU56LMIPRX3YUWV2LAVCNFSM6AAAAABCX66XACVHI2DSMVQWIX3LMV45UABCJFZXG5LFIV3GK3TUJZXXI2LGNFRWC5DJN5XDWMJRHA3TENRSGUZTMMQ>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>